% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternal_mortality.R
\docType{data}
\name{maternal_mortality}
\alias{maternal_mortality}
\title{Maternal Mortality Data}
\format{
A data frame with 10 rows and 23 variables:
\describe{
  \item{iso}{Country ISO code}
  \item{country}{Country name}
  \item{indicator}{Indicator name}
  \item{X2000 ... X2019}{Mortality values for each year}
}
}
\source{
World Bank
}
\usage{
maternal_mortality
}
\description{
A small sample dataset in World Bank format (years as wide columns).
}
\keyword{datasets}
